Fix: use a dedicated model for instance rootfs
Problem: re-using the persistent volume model is not a perfect fit
when implementing the logic to support instances in the rest of
the aleph.im stack. Ex: the `mount` point does not make sense
(it will always be `/`), neither do the `comment` or `name` fields.
Similarly, the `parent` field must always be specified for root
file systems.

Solution: use a custom model to refine the fields required.