Change: Message.time was float, becomes datetime (#40)
* Change: Message.time was float, becomes datetime

Change the type if BaseMessage.time to a datetime. This allows easier manipulation of the field, and parsing from strings as well as floats.

New APIs introduced in PyAleph 0.5.0 store the message time as timestampz and return it as ISO strings. This changes is compatible with both the old and new APIs.

* Add time validator for auto-conversion from float to datetime

* Fix missing parameter

* Fix Dockerfile dependency installation

* Fix typing issues with mypy by using a generic type for AlephMessage

* Rename to convert_float_to_datetime

Co-authored-by: Hugo Herter <git@hugoherter.com>

---------

Co-authored-by: mhh <mike.hukiewitz@robotter.ai>
Co-authored-by: Mike Hukiewitz <70762838+MHHukiewitz@users.noreply.github.com>